dsa/dsv4: widen token_id in the remaining k-cache dequant kernels
Same int32-overflow vector as the quant side: token_id * output_stride_0
and token_id_paged * pool strides are int32 multiplies that wrap on large
pools before widening. Cover the paged dequant kernel and the dsv4
dequant kernel so the whole fp8 k-cache quant/dequant family is int64.

+++ python/sglang/kernels/ops/attention/dsa/dequant_k_cache.py
@@ -247,8 +247,8 @@ def _dequantize_k_cache_paged_kernel(
     DIM_NOPE: tl.constexpr,
     DIM_ROPE: tl.constexpr,
 ):
-    token_id = tl.program_id(0)
-    token_id_paged = tl.load(page_table_1_ptr + token_id).to(tl.int32)
+    token_id = tl.program_id(0).to(tl.int64)
+    token_id_paged = tl.load(page_table_1_ptr + token_id).to(tl.int64)
     raw_block_id = tl.program_id(1)
 
     if raw_block_id < NUM_NOPE_BLOCKS:
